server/requirements.txt
# FastAPI and server
fastapi==0.115.6
uvicorn[standard]==0.34.0
python-multipart==0.0.20

# Telegram Bot
python-telegram-bot==21.9

# Supabase
supabase==2.11.0

# JWT and security
PyJWT==2.10.1
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
cryptography==44.0.0

# Environment and config
python-dotenv==1.0.1
pydantic>=2.10.5
pydantic-settings==2.7.1

# HTTP client
httpx==0.28.1
